+++ src/antpack/cpp_src/utilities/dna_sequence_utilities.cpp
@@ -27,9 +27,7 @@
 namespace DNASequenceUtilities {
 
 
-/// @brief Gets the reverse complement letter for an input DNA letter.
-/// @param letter The input letter.
-/// @return The output reverse complement letter.
+
 char reverse_complement_letter(const char &letter) {
     switch (letter) {
         case 'A':
@@ -59,16 +57,11 @@ char reverse_complement_letter(const char &letter) {
 
 
 
-/// @brief A hash function for DNA codons so that each is mapped to a unique
-/// value.
-/// @param codon The DNA codon.
-/// @return An integer. All possible DNA codons will have a unique integer
-/// assigned.
 constexpr int64_t codon_hash(const char *codon) {
     const int64_t p = 131;
     int64_t total = 0;
     int64_t multiplier = 1;
-    for (size_t i=0; codon[i] != '\0'; i++) {
+    for (int i=0; codon[i] != '\0'; i++) {
         total += multiplier * codon[i];
         multiplier *= p;
     }
@@ -77,10 +70,7 @@ constexpr int64_t codon_hash(const char *codon) {
 
 
 
-/// @brief Converts an input DNA codon to an amino acid. Throws an exception
-/// which can be handled by Python caller if unrecognized codon is supplied.
-/// @param codon The DNA codon.
-/// @return An amino acid letter.
+
 char codon_to_aa(const char *codon) {
     switch (codon_hash(codon)) {
         case codon_hash("TTT"):
